nested interval progress

+++ Laravel/database/migrations/2017_03_24_233441_create_groups_table.php
@@ -40,7 +40,7 @@ class CreateGroupsTable extends Migration
                 ->nullable()
                 ->unsigned();
 
-            $table->integer('interval')
+            $table->bigInteger('interval')
                 ->nullable()
                 ->unsigned()
                 ->index();
